Implemented array style configuration loading, PriorityList data structure.
<?php
return array('owner' => 'John Doe';);
?>

Files MUST end with .php.

Sorted Array (will be sorted on every insert). When running a foreach, the element with the highest
priority will be read first and so on. Elements with the same priority can be inserted side
by side, but the first one inserted has a higher priority. Values don't need to be unique.

This class is used in ConfigFile to provide an interface similar to spl_autoload_register, but with
the possibility to change the order through values.